Change JSON format
The original format was quick-n-dirty and more like a "raw dump" of the
compiler configurations. The quick-n-dirty format would probably work
just find for 99% of the use cases but it was kind of noisy and verbose.
The target features were not great. The new format is less noisy (no
"target_" everywhere) and provides an array of the target features. It
is just more clean and elegant.

Note, there is no implementation or CLI argument to select a JSON format
version. I am not sure such an implementation is even needed.