Also emit source information for paths
Previously, the location for path-based (i.e., local) packages was
redacted from metadata output. This was done way back when so that
`Cargo.lock` would not include path sources:
https://github.com/rust-lang/cargo/issues/7483#issuecomment-539258083

This PR makes this redaction optional so that `cargo metadata` and
friends _will_ output the source location, which is useful for tools
like `rustfmt` which need to know the paths to local packages (see #7483
and rust-lang/rustfmt#4599).

Interestingly enough, no tests fail even though
`disable_path_serialization` is never currently called. I don't know if
that's because there's no test that cover where redaction is needed, or
because redaction is simply no longer needed (presumably because
lockfile generation now uses custom logic). If it is the latter, this PR
can be simplified to not support opting into redaction.

Fixes #7483