Switch to v2 API url for new feature syntax.
This switches the API url for publishing from `api/v1/crates/new` to
`api/v2/crates/new` if the package contains the new features syntax.

This is necessary because if the registry does not know about the new
feature syntax, it will (likely) silently accept the uploaded package
without the `features2` field, which will end up creating an index
with an invalid entry (missing features).