Add a schema version to the index.
This adds a `v` field which indicates a format version for an index
entry. If Cargo encounters a version newer than it understands,
it will ignore those entries. This makes it safer to make changes
to the index entries (such as adding new things), and not need to
worry about how older cargos will react to it.

This version field is also part of the publish API.

Version `2` indicates that it uses the new features field.