Remove `--print-cfg` argument
The `--print-cfg` flag will be replaced with the `--print foo` option
and passed to `rustc --print foo`. The output will be the "raw" output
from the `rustc --print foo` command instead of a serialized JSON of the
RustcTarget Data. This is the first step in changing the implementation.
The serialization code is no longer needed.