Improve error messages for wrong target paths
As it was stated in #9014 it's not really useful the message
that you get when using `bench/your_bench.rs` instead of the
correct path `benches/your_bench.rs`.

Therefore, as @ehuss suggested, the error message has been improved
not only suggesting the correct path but also suggesting to
specify `bench.path` in `Cargo.toml` for non-default path usage.
And this has been exented to `test`, `example` and `bin`.