Allow setting cargo home from .cargo/home
This patch allows users to set the Cargo home directory through the file
system using a pointer file in `.cargo/home`. As with `.cargo/config`,
the file is searched for recursively up the tree from the current
working directory. Unlike with `.cargo/config`, the search terminates
when a matching file is found.

The search for `.cargo/home` happens before any `.cargo/config` files
are read, as that may be affected by the choice of Cargo home directory.

The `CARGO_HOME` environment variable is preferred over `.cargo/home`.

Fixes #6452.