Add support for preview features
This is an experimental implementation of a feature that adds _preview
features_ to Cargo. A preview feature is one that is still considered
unstable, but is available on stable/beta for testing behind a new flag,
`--enable-preview` that requires you explicitly enumerate the features
to enable.

The core idea is that the Cargo team would specifically designate
a small set (potentially zero at any given point in time) of features as
"available for preview", subject to criteria that ensure that a preview
period of a feature is both worthwile and won't cause undue breakage. A
rough framework for how to think of these was proposed in
https://internals.rust-lang.org/t/survey-of-high-impact-features/14536.

This implementation exists mainly to get feedback on whether the
approach taken here is reasonable from a code point-of-view. Before this
feature lands (if it does), it would have to go through an RFC process.

For the initial implementation, I have designated one unstable feature
(`patch-in-config`) and one unstable option (`--out-dir`) as "preview".
This choice was based on some of the internals discussion linked above,
but is in no way final.