cargo-run-deps: Make commands in dependencies available to run
This approach teaches `cargo run` to handle also binary
targets in dependencies. In turn, it allows using cargo as a shim
when needing to invoke a specific package version from the
set of dependencies, without resorting to install a user-wide
version of a binary.

Users of e.g. `mdbook` should rejoice.

Fixes #2267.