feat: display warning if target/doc contains unexpected docs
This commit modifies the behavior of cargo doc to warn the user if the
output directory for the documentation includes rendered docs for target
units/crates which were not compiled as a part of this invocation.
The implementation simply checks for entries in the target/doc/src
directory whose name contains an unexpected unit/crate name.

This change guides users on how to correct the rendered documentation if
previous invocations of the command caused undesired crates to be
documented. For example, were a user to invoke the command to document
all workspace crates and dependencies and then re-run the command using
the --no-deps flag, the rendered documentation will still include docs
for all dependencies. Clearly, the solution is to `rm -rf target/docs`
and re-run using the added flag, but this may not be obvious to all
users, and a helpful warning will give them a nudge in the right
direction.

Signed-off-by: Sean Pianka <pianka@eml.cc>