Platform-specific-ness is carried via deps
If e.g. an optional feature is only enabled on certain platforms,
meaning a transitive dependency is only enabled on certain platforms,
currently that information can't be worked out from the output of
`cargo metadata`.

With this change, when a dependency is configured for a specific
platform, its dependencies are by default also configured for that
specific platform, unless they explicitly have a different platform
attached to them.

This makes it easier to translate crates.io packages for building with
non-cargo build systems, because this platform-specific behaviour of
feature resolution is now observable in `cargo metadata`, rather than
requiring those systems to track and re-implement feature resolution
themselves with regard to platforms.

Fixes https://github.com/rust-lang/cargo/issues/9863.