clean: add --include-cache option
This patch adds the `--include-cache` flag to `cargo clean`, which
allows the command to also remove related artifacts from `CARGO_HOME`
such as downloaded `.crate` files in `~/.cargo/cache/` and extracted
source directories in `~/.cargo/src/`.

Note that this feature is not intended to replace the `cargo-cache`
command which does smart cache management. Instead, this command simply
blows away whatever it's asked to delete with no smarts whatsoever.

Fixes #3289.