Clarify configuration and workspace interaction
The way it is currently phrased is not actually consistent with the behaviour I experienced, in which even I specify the `--manifest-path` flag value, the config value gets ignored if the target crate is in a workspace, regardless of whether I am invoking it from the workspace root. What counts is the crate location w.r.t. the workspace, not where cargo is invoked from (I think). Please feel free to double check my understanding.