Add 'multidep' feature toggle, initial impl of RFC-3176
This is a squashed commit comprised of the following:
-----------------------------------------------------

Add metadata test to see how it would respond to multiple names for a crate with the same version

Add all multidep specific artifact tests

They are ignored for now until the actual implementation comes in for
validation.

Add support for multiple names for the same crate

It's worth mentioning that crates, called packages, are identified
by their name and their semver version. The resolver handles
figuring out which versions are allowed, but typically separates
crates of the same name by breaking semver version. Thus a:0.1 and
a:0.2 are treated as two different dependencies as far as the code
is concerned. a:0.1.0 is a package id.

In order to allow multiple names to the same package id code that used
to be fine with a single id now has to deal with multiple pairs of
package-id/dependency-name pairs.

The key to understanding why so many functions are affected is
Resolver::extern_crate_names_and_dep_names() which trickles up
into unit-dependencies.

Re-activate all tests validating 'multidep' thus far

Remove adjustments to `cargo metadata` test with bindeps

It's there to show what currently happens if there are multideps,
which basically means these are ignored entirely, showing only
the first available name of the package.

Fix build errors due to dep-filter signature changes