Don't fetch from the remote repo if the revision is already present locally
This makes `cargo update` *much* faster when there are many git deps in a repo and they're already
cached locally.

There was an existing check for this, but it didn't quite hit all cases.
The code looked like this:
```rust
        let (db, actual_rev) = match (self.locked_rev, db) {
            (Some(rev), Some(db)) if db.contains(rev) => (db, rev),
            (None, Some(db)) if self.config.offline() => {
                let rev = db.resolve(&self.manifest_reference)?;
                (db, rev)
            }
            (locked_rev, db) => self.remote.checkout(),
        };
```

The problem was that when `locked_rev.is_none() && db.is_some()` and `--offline` wasn't passed,
cargo would skip the `db.resolve()` check, and issue a fetch from the remote even if it wasn't
necessary.

This changes the `db.resolve()` check to happen even in online mode.
This brings `cargo update` down from 53 to 6 seconds for me on a codebase with 10+ git dependencies.