[beta] Remove `strip = "off"` (and undocumented `strip = "n"`/`strip = "no"`)
The `strip` option allows the string `"off"`, which `rustc` doesn't
have, and maps it to `"none"`. Remove that equivalent, and just keep
`"none"`.

Simplify and clarify the documentation for `strip` accordingly.

`strip` reused the `is_off` function from `lto`, which also had the net
effect of allowing `"n"` and `"no"`, which weren't documented as
supported. This change drops those as well.