Expose finer-grained resolver machinery
Previously, users consuming Cargo as a library had to either use the
very low-level `resolve_with_previous`, or the "do everything"
`resolve_ws_with_opts`, which includes downloading the source tarballs
of most crates in the resolved dependency closure. This patch exposes a
version of `resolve_ws_with_opts` that does everything _up to_ anything
that requires a source tarball download, as well as `resolve_registry`,
which is a building block of other resolver functions that may be of
use.