Handle Windows-reserved paths on all platforms
Cargo had error handling and test coverage for the fact that path names
such as `aux.rs` are not allowed on Windows. More recent versions of
Windows do support such path names, but given that Cargo still supports
older versions it would be better to have consistent handling of path
names across all platforms. Indeed, there is already code to prevent
packages containing reserved path names from being published, so it
makes sense to be consistent elsewhere.

This commit changes the behavior to be consistent and tested on all
platforms that Cargo supports.