Completion support for `cargo-add`
cargo-add's PR was minimal to ensure we had settled on the CLI before
adding references to the CLI and have cascading churn.  This updates
bash and zsh completion to include cargo-add.

I am unsure how to handle testing of completions, so this is a
best-effort from looking at other completions and the docs.