This change adds an example to the authors attribute in the manifest.
It is submitted to clarify the documenation for new folks such as
myself. See the forum issue linked below which prompted this change
since the error message the compiler issues is not clear for new
folks.

https://users.rust-lang.org/t/compile-error-when-adding-authors-to-cargo-toml/79383

Thank you for considering this contribution. It's my first so go
easy on me please!  :-)