refactor(cli): Migrate to explicit ArgAction
This is prep for migrating to clap v4 where `ArgAction::IncOccurrences` and
`ArgAction::StoreValue` (the implicit ones) are being removed.  This
will also prevent us from hitting issues from `Arg::new().long()` now
defaulting to taking a value.

The main side effect from this is that `ArgAction::Set` overrides
previous occurrences, rather than conflicting with them.  This removes a
couple of cases we explicitly test for.