Auto merge of #11610 - ehuss:disable-public-network-windows, r=epage
Disable network SSH tests on windows.

These tests have a high failure rate on Windows, so this disables them for now. I don't know exactly why they are failing. If I had to take a wild guess, I would suspect the use of WinCNG for the SSH backend, which may mean it is out of our control.