assure broken symlinks don't stop walkdir when cleaning target dirs
This issue was encountered in the `git::recompilation()` test which
failed as `foo.dSYM` was a link whose target was deleted already.

When passing it to `WalkDir` as root, the expectation is that it is
returned as entry, even if it's not a directory, from where it can
then be deleted.

Even though `follow_symlinks` is off by default, when the passed in
root (which is assumed to be a directory) is a symlink, it tries
to obtain its meta-data, and following this broken symlink fails
which causes the only returned entry to be an error.

Now we check for this case, which should cost us only in the few cases
where files are symlinks, and remove the symlink specifically.

I thought about raising this as issue for `WalkDir` but thought that
given the documentation for [WalkDir::new(...)](https://docs.rs/walkdir/2.3.2/walkdir/struct.WalkDir.html#method.new)
this behaviour seems to be intended.