Auto merge of #11661 - arlosi:no-error-for-auth-required, r=Eh2406
Do not error for `auth-required: true` without `-Z sparse-registry`

Registries that include `auth-required: true` in their `config.json` currently hit the following error on stable:
```
authenticated registries require `-Z registry-auth`
```

This situation makes it difficult for a registry to optionally offer the `auth-required: true` feature, since it forces users on to the nightly toolchain.

This PR changes the behavior to ignore the `auth-required: true` field of `config.json` without `-Z registry-auth`.

The downside to this change is that it makes it harder to discover why a registry isn't working, since the user will get an HTTP 401 error from the server, rather than a message from Cargo suggesting adding `-Z registry-auth`.

r? `@Eh2406`