docs(contrib): Move Design Principles earlier in the book
This is the framing by which all contributions should happen, so this
should come first before getting into what is needed to know to
implement a change and test it.

This is also prep for adding a new Implementations Practices section to
pull out various tips spread through the Architecture section.  See #11841