chore: Use globs for workspace members
This is a short-term option until we can have a better solution for
globbing.  This does not update `benches/` to support which has a README
in there preventing globbing; this seems low-churn enough not to find a
solution for it.

On the next sync-up with rust-lang/rust, we'll need to update https://github.com/rust-lang/rust/blob/4e463012580415a932ae4fc255aff45982c70369/src/bootstrap/tool.rs#L588-L603

Fixes #11988