refactor: use the term "index schema" instead of "index format"
We had index format version and index file version. Both sometimes
are written as "index version". This is so confusing that I need to
check to source code back-and-forth.

Today we make "index format version", which refers to the exact
representation of a index file, rename to "index schema version".
Other index file version of index version should just mean the opaque
version of an index file for cache invalidation use.