Allow the edition specifier to be an integer
This helps save the number of characters needed to specify an edition.
So far, each edition was fully described by its year number, 2015, 2018, 2021.
Also for the future, the intent is that editions will always just be
year numbers. Therefore, we don't have to use a toml string for them in Cargo.toml,
an integer is enough. This saves two characters in each Cargo.toml.