chore: Automatically update dependencies monthly
This will update `Cargo.toml` and `Cargo.lock` at the same time.  By
default, all dependencies will be updated, so we pull out the compatible
dependencies into a consolidated PR.

This starts off with the dependency dashboard and automerging being
disabled but explicitly specified to call attention to the potential for
this.  We can evaluate these later.