docs(ci): Remove Travis example
When the guide was written, it only included Travis CI.  For while, it
was *the* CI service people used.  However, since git hosting services
have been integrating CI support and with Travis CI's new owners, market
share has been going down.  One estimate I found said that Travis CI's
market share is 0.94%.  If we had a compelling reason to include it,
independent of that, I would.  However, including it and making it first
in the list is a distraction.  There are CI services in more common use
(e.g. CircleCI, Jenkins) to include at this point over Travis.