docs: Generalize `Cargo.lock` guidance
Before, we were fairly prescriptive of when to check-in the `Cargo.lock`
file.   The guidance has been updated to instead explain the trade offs
and to mostly leave it in users hands.  As a starting point, we do say
to check it in for those that want an "easy answer".