chore: Downgrade serde below the binary blob
As of serde 1.0.172, `serde_derive` ships a binary blog for Linux x64
for faster build times.  This blob is not yet reproducible to ensure
that the safety of it.  See serde-rs/serde#2538

This is not a judgement on serde or on dtolnay but just a precaution to
buy us more time as the community works through this since the beta cut
is coming up.  rust-1.72 branch is unaffected.