refactor: Use more serde_untagged
I felt this does a good job of cleaning up the code and by using it
more, new uses are more likely to use it.

Due to an error reporting limitation in `serde_untagged`, I'm not using
this for some `MaybeWorkspace` types because it makes the errors worse.

I also held off on some config visitors because they seemed more
complicated and I didn't want to risk that code.