Wrap clean operations in a CleanContext.
This refactors some of the `cargo clean` code to wrap the "cleaning"
operation in a `CleanContext` so that the context can be passed to other
parts of cargo which can perform their own cleaning operations.

There are some minor changes in the error messages to prepare for
cleaning operations that aren't directly related to the build directory.