refactor(toml): Move impl bodies next to types
At minimum, we should have "everything" for a type next to it but it was
previously scattered.

When collecting the related impl's, I decided to order based on how I
tend to do it.  Generally the methods are the "core" of the type so I
put those first.

The order I went with
1. Basic impl body
2. Complex impl body
3. Trait impls
4. Helper free-functions