Rename some `Filesystem` methods for clarity and consistency.
This tries to align the `Filesystem` method names so they have the same
form, and also so they spell out exactly what they do. (An alternative
would be to have a builder, similar to `OpenOptions`, but that seems a
bit overkill here.)

This also extends the docs to try to make it clearer how this works.