cargo/artifact: prepare compatibility env-vars
We are about to change the default value for target-names of libraries.
They used to match the package-name. In the future, they will use the
package-name with dashes converted to underscores. This will affect the
artifact env-variables, since they expose target-names. Hence, set the
old env-vars, too, to avoid breakage.

Note that we do not retain the name of a target before it was converted,
and the conversion is lossy, so we cannot reconstruct it. However, we
can rely on the fact that the conversion only happens for default values
(since user-supplied values never allowed dashes). Furthermore, we now
remember whether a target-name was inferred, so we can exactly
reconstruct whether a library-target could have contained dashes in
older releases, or not.