requirements from replacements do not come from the lock file
I think the core of the problem is that https://github.com/rust-lang/cargo/commit/725420e8d788e003ee82c15f66d50272aab493d1 was a mistake. Before that PR we used `=` constraints to imply that it came from lock file so when we saw a `=` constraints being constructed for a replace it should use the new system for marking something as coming from lock file. But I don't think that is the semantics that an `=` constraint implies in this context.